0.2023.09.11
Bug fix where method color() was not properly being accessed as self.color()
Change Toggle Help button in wire editor dialog to a Show help checkbox
Add a checkbox to hide ghost object in wire edit dialog
Add a checkbox to hide WireFilter object in wire edit dialot
Check if WireFilter was made with preselected subelements (faces or edges of source object) and do not allow wire editing if so, since it doesn't work properly in this situations.  Instead, the user is advised to create a new WireFilter of this WireFilter and edit that wire order instead.